+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Utils.Str.getCopyright()
 - Some Guard and Assert isNumberX() functions.
 - Utils.Str.splitByCapsAndSpaces()
+- Utils.Str.splitByStrings()
+- Utils.Str.splitByChars()
 
 ## [2.1.0] - 2025-11-01
 ### Added